(maint) Scrape peconf for console password
If the peconf is set to a password that is not the default beaker console password (i.e. frankenbuilds), password verification fails and the run hangs in this state. This was seen attempting to create a pe_compiler node through a franeknbuild, where `run_puppet_access_login` is called with default values instead od frankenbuilder's admin pass.
This pr scrapes peconf for a password and replaces the default is there is one available.